remove old prepro script and related code
